Today's work on the code and report. I have done little changes on the code. I have ran it some times and it seems to work well. The final errors on the parameters increase if we increase the steps hEpsion and hTemp...
It seems that the He10 data corresponds to superconductor-superconductor junction... It's amazing! The big-little mountain inside the gap is too near to the value (gap_Pb - gap_Al), which is the point where must appear a mountain in the S-S junction. I guess we have got an S-S junction...

I have worked a little bit on the report. I will continue those days expanding the part of the theory... so you can do another part if you want. 

Enjoy your feiertagssssssss?? (it's correct?? :-) )