New improvements...
I still think that our best data sample corresponds to superconductor-superconductor junction. 

I have made the computation for some different data files, and the Temperature is different but the gap and conductance is nearly equal, as expected.

Increasing the hTemp and hEpsilon steps, we get greater covar values... But I think this is not necessary, the algorithm is correct and the parameters given are reasonably right...